+++ src/bsvar_t.cpp
@@ -53,23 +53,24 @@ Rcpp::List bsvar_t_cpp(
   mat     aux_B       = as<mat>(starting_values["B"]);
   mat     aux_A       = as<mat>(starting_values["A"]);
   mat     aux_hyper   = as<mat>(starting_values["hyper"]);
-  vec     aux_lambda  = as<vec>(starting_values["lambda"]);
-  double  aux_df      = as<double>(starting_values["df"]);
+  mat     aux_lambda  = as<mat>(starting_values["lambda"]);
+  vec     aux_df      = as<vec>(starting_values["df"]);
   
   const int   SS    = floor(S / thin);
   
   cube  posterior_B(N, N, SS);
   cube  posterior_A(N, K, SS);
   cube  posterior_hyper(2 * N + 1, 2, SS);
-  mat   posterior_lambda(T, SS);
-  vec   posterior_df(SS);
+  cube  posterior_lambda(N, T, SS);
+  mat   posterior_df(N, SS);
   mat   tmp_lambda_sqrt(N, T);
   
   int   ss = 0;
   
   // the initial value for the adaptive_scale is set to the negative inverse of 
   // Hessian for the posterior log_kenel for df evaluated at df = 30
-  double adaptive_scale = pow(0.25 * T * R::psigamma(15, 1) - T * pow(17, -2) - 2 * pow(16, -2), -1);
+  double  adaptive_scale_init = pow(0.25 * T * R::psigamma(15, 1) - T * pow(17, -2) - 2 * pow(16, -2), -1);
+  vec     adaptive_scale(N, fill::value(adaptive_scale_init));
   
   for (int s=0; s<S; s++) {
   
@@ -78,23 +79,24 @@ Rcpp::List bsvar_t_cpp(
     // Check for user interrupts
     if (s % 200 == 0) checkUserInterrupt();
     
-    vec df_tmp      = sample_df ( aux_df, adaptive_scale, aux_lambda, s, adptive_alpha_gamma );
-    aux_df          = df_tmp(0); 
-    adaptive_scale  = df_tmp(1);
+    List df_tmp      = sample_df ( aux_df, adaptive_scale, aux_lambda, s, adptive_alpha_gamma );
+    
+    aux_df          = as<vec>(df_tmp["aux_df"]);
+    adaptive_scale  = as<vec>(df_tmp["adaptive_scale"]);
     
     aux_lambda      = sample_lambda ( aux_df, aux_B, aux_A, Y, X );
-    tmp_lambda_sqrt.each_row() = sqrt(aux_lambda.t());
+    tmp_lambda_sqrt = sqrt(aux_lambda);
     
     aux_hyper       = sample_hyperparameters(aux_hyper, aux_B, aux_A, VB, prior);
     aux_A           = sample_A_heterosk1 ( aux_A, aux_B, aux_hyper, tmp_lambda_sqrt, Y, X, prior);
     aux_B           = sample_B_heterosk1 ( aux_B, aux_A, aux_hyper, tmp_lambda_sqrt, Y, X, prior, VB );
     
     if (s % thin == 0) {
-      posterior_B.slice(ss)     = aux_B;
-      posterior_A.slice(ss)     = aux_A;
-      posterior_hyper.slice(ss) = aux_hyper;
-      posterior_lambda.col(ss)  = aux_lambda;
-      posterior_df(ss)          = aux_df;
+      posterior_B.slice(ss)       = aux_B;
+      posterior_A.slice(ss)       = aux_A;
+      posterior_hyper.slice(ss)   = aux_hyper;
+      posterior_lambda.slice(ss)  = aux_lambda;
+      posterior_df.col(ss)        = aux_df;
       ss++;
     }
   } // END s loop