+++ src/sample_t.cpp
@@ -8,8 +8,8 @@ using namespace arma;
 
 // [[Rcpp::interfaces(cpp)]]
 // [[Rcpp::export]]
-arma::vec sample_lambda (
-    const double&       aux_df,
+arma::mat sample_lambda (
+    const arma::vec&    aux_df,     // Nx1
     const arma::mat&    aux_B,      // NxN
     const arma::mat&    aux_A,      // NxK
     const arma::mat&    Y,          // NxT
@@ -19,9 +19,16 @@ arma::vec sample_lambda (
   const int T           = Y.n_cols;
   
   mat       U           = aux_B * ( Y - aux_A * X );
-  vec       s_lambda    = aux_df + 2 + trans(sum( pow(U, 2) ));
-  double    nu_lambda   = aux_df + N;
-  vec       aux_lambda  = s_lambda / as<vec>(Rcpp::rchisq(T, nu_lambda));
+  
+  mat       s_lambda    = pow(U, 2);
+  s_lambda.each_col()  += aux_df + 2;
+  vec       nu_lambda   = aux_df + N;
+  
+  mat       aux_lambda(N, T);
+  for (int n=0; n<N; n++) {
+    vec draw            = as<vec>(Rcpp::rchisq(T, nu_lambda(n)));
+    aux_lambda.row(n)   = s_lambda.row(n) / draw.t();
+  }
   
   return aux_lambda;
 } // END sample_lambda
@@ -31,7 +38,7 @@ arma::vec sample_lambda (
 // [[Rcpp::export]]
 double log_kernel_df (
     const double&       aux_df,
-    const arma::vec&    aux_lambda  // Tx1
+    const arma::rowvec&    aux_lambda  // Tx1
 ) {
   
   const int T   = aux_lambda.n_elem;
@@ -48,30 +55,34 @@ double log_kernel_df (
 
 // [[Rcpp::interfaces(cpp)]]
 // [[Rcpp::export]]
-arma::vec sample_df (
-    double&           aux_df,
-    double&           adaptive_scale,
-    const arma::vec&  aux_lambda,         // Tx1
+Rcpp::List sample_df (
+    arma::vec&        aux_df,             // Nx1
+    arma::vec&        adaptive_scale,     // Nx1
+    const arma::mat&  aux_lambda,         // NxT
     const int&        s,                  // MCMC iteration
     const arma::vec&  adptive_alpha_gamma // 2x1 vector with target acceptance rate and step size
 ) {
+  int N = aux_df.n_elem;
+  vec aux_df_star(N);
+  vec alpha(N, fill::ones);
   
   // by sampling from truncated normal it is assumed that the asymmetry from truncation 
   // is negligible for alpha computation
-  double aux_df_star  = RcppTN::rtn1( aux_df, adaptive_scale, 0, R_PosInf );
-  
-  double alpha        = 1;
-  double kernel_ratio = exp( log_kernel_df(aux_df_star, aux_lambda) - log_kernel_df(aux_df, aux_lambda) );
-  if ( kernel_ratio < 1 ) alpha = kernel_ratio;
-  
-  if ( R::runif(0, 1) < alpha ) {
-    aux_df = aux_df_star;
-  }
-  
-  if (s > 1) {
-    adaptive_scale = exp( log(adaptive_scale) + 0.5 * log( 1 + pow(s, - adptive_alpha_gamma(1)) * (alpha - adptive_alpha_gamma(0))) );
-  }
-  
-  vec out = {aux_df, adaptive_scale};
-  return out;
+  for (int n = 0; n < N; n++){
+    aux_df_star(n)        = RcppTN::rtn1( aux_df(n), adaptive_scale(n), 0, R_PosInf );
+    double kernel_ratio   = exp( log_kernel_df(aux_df_star(n), aux_lambda.row(n)) - log_kernel_df(aux_df(n), aux_lambda.row(n)) );
+    
+    if ( kernel_ratio < 1 ) alpha(n) = kernel_ratio;
+    if ( R::runif(0, 1) < alpha(n) ) {
+      aux_df(n) = aux_df_star(n);
+    }
+    if (s > 1) {
+      adaptive_scale(n) = exp( log(adaptive_scale(n)) + 0.5 * log( 1 + pow(s, - adptive_alpha_gamma(1)) * (alpha(n) - adptive_alpha_gamma(0))) );
+    }
+  } // END n loop
+
+  return List::create(
+    _["aux_df"] = aux_df,
+    _["adaptive_scale"] = adaptive_scale
+  );
 } // END sample_df